NAS-105409 / 11.3 / Implement proper way of checking if destination has data (by themylogin)

+++ integration-tests/replication/test_target_without_snapshots_but_with_data.py
@@ -10,7 +10,12 @@ from zettarepl.utils.test import run_replication_test
 
 
 @pytest.mark.parametrize("zvol", [False, True])
-def test_target_without_snapshots_but_with_data(zvol):
+@pytest.mark.parametrize("mounted", [False, True])
+@pytest.mark.parametrize("snapdir", [False, True])
+def test_target_without_snapshots_but_with_data(zvol, mounted, snapdir):
+    if zvol and (not mounted or snapdir):
+        return
+
     subprocess.call("zfs destroy -r data/src", shell=True)
     subprocess.call("zfs receive -A data/dst", shell=True)
     subprocess.call("zfs destroy -r data/dst", shell=True)
@@ -23,11 +28,19 @@ def test_target_without_snapshots_but_with_data(zvol):
     subprocess.check_call("zfs snapshot -r data/src@2018-10-01_02-00", shell=True)
 
     if zvol:
-        subprocess.check_call("zfs create -V 10m data/dst", shell=True)
-        subprocess.check_call("dd if=/dev/urandom of=/dev/zvol/data/dst bs=1m count=1", shell=True)
+        subprocess.check_call("zfs create -V 20m data/dst", shell=True)
+        subprocess.check_call("dd if=/dev/urandom of=/dev/zvol/data/dst bs=15m count=1", shell=True)
     else:
         subprocess.check_call("zfs create data/dst", shell=True)
-        subprocess.check_call("dd if=/dev/urandom of=/mnt/data/dst/test bs=1m count=1", shell=True)
+        if snapdir:
+            subprocess.check_call("zfs set snapdir=visible data/dst", shell=True)
+        if mounted:
+            bs = "1k"
+        else:
+            bs = "15m"
+        subprocess.check_call(f"dd if=/dev/urandom of=/mnt/data/dst/test bs={bs} count=1", shell=True)
+        if not mounted:
+            subprocess.check_call("zfs unmount data/dst", shell=True)
     time.sleep(5)  # "used" property is not updated immediately
 
     definition = yaml.safe_load(textwrap.dedent("""\
